% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utils.R
\name{.ds_classes}
\alias{.ds_classes}
\title{Internal: Combine class names}
\usage{
.ds_classes(...)
}
\arguments{
\item{...}{Class names (can be NULL)}
}
\description{
Internal: Combine class names
}
\keyword{internal}
