% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-card.R
\name{ds_card_block}
\alias{ds_card_block}
\title{Card Block}
\usage{
ds_card_block(..., class = NULL)
}
\arguments{
\item{...}{Child elements}

\item{class}{Additional CSS classes}
}
\value{
A Shiny tag object
}
\description{
A block section within a card with proper padding.
}
