% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-input.R
\name{ds_checkbox}
\alias{ds_checkbox}
\title{Checkbox Component}
\usage{
ds_checkbox(inputId, label, value = FALSE, size = NULL, disabled = FALSE, ...)
}
\arguments{
\item{inputId}{The input slot for Shiny reactivity}

\item{label}{The checkbox label}

\item{value}{Initial checked state}

\item{size}{Size variant ("sm", "md", "lg")}

\item{disabled}{If TRUE, disables the checkbox}

\item{...}{Additional attributes}
}
\value{
A Shiny tag object
}
\description{
Create a styled checkbox using Designsystemet styles.
}
\examples{
ds_checkbox("agree", "I agree to the terms", value = FALSE)
}
