% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-misc.R
\name{ds_chip_group}
\alias{ds_chip_group}
\title{Chip Group Component}
\usage{
ds_chip_group(inputId, ..., class = NULL)
}
\arguments{
\item{inputId}{Shiny input ID}

\item{...}{\code{ds_chip()} elements}

\item{class}{Additional CSS classes}
}
\value{
A Shiny tag object
}
\description{
Create a reactive group of toggle chips. Selected chip values are reported
to Shiny as a character vector via \code{input[[inputId]]}.
}
\examples{
\dontrun{
ds_chip_group(
  "languages",
  ds_chip("R",      value = "r",      selected = TRUE),
  ds_chip("Python", value = "python"),
  ds_chip("Julia",  value = "julia")
)
}
}
