% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-dependencies.R
\name{ds_dependencies}
\alias{ds_dependencies}
\title{Designsystemet Dependencies}
\usage{
ds_dependencies()
}
\value{
A list of htmltools htmlDependency objects
}
\description{
Returns the HTML dependencies required for Designsystemet components.
}
\examples{
ds_dependencies()
}
