% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-misc.R
\name{ds_divider}
\alias{ds_divider}
\title{Divider Component}
\usage{
ds_divider(class = NULL)
}
\arguments{
\item{class}{Additional CSS classes}
}
\value{
A Shiny tag object
}
\description{
Create a horizontal divider using Designsystemet styles.
}
\examples{
ds_divider()
}
