% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-error-summary.R
\name{ds_error_summary}
\alias{ds_error_summary}
\title{Error Summary Component}
\usage{
ds_error_summary(..., heading = "Feil i skjema", class = NULL)
}
\arguments{
\item{...}{Error list items}

\item{heading}{The heading text for the error summary}

\item{class}{Additional CSS classes}
}
\value{
A Shiny tag object
}
\description{
Display a summary of form validation errors.
}
\examples{
ds_error_summary(
  heading = "There were errors in the form",
  htmltools::tags$li(ds_link("Name is required", href = "#name")),
  htmltools::tags$li(ds_link("Email is invalid", href = "#email"))
)
}
