% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-typography.R
\name{ds_label}
\alias{ds_label}
\title{Label Component}
\usage{
ds_label(..., `for` = NULL, class = NULL)
}
\arguments{
\item{...}{Label content}

\item{for}{The ID of the input this label is for}

\item{class}{Additional CSS classes}
}
\value{
A Shiny tag object
}
\description{
Create a styled form label using Designsystemet styles.
}
\examples{
ds_label("Email address", `for` = "email")
}
