% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-list.R
\name{ds_list_item}
\alias{ds_list_item}
\title{List Item}
\usage{
ds_list_item(...)
}
\arguments{
\item{...}{Item content}
}
\value{
A Shiny tag object
}
\description{
Create a list item.
}
\examples{
ds_list_item("First item")
}
