% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-misc.R
\name{ds_popover}
\alias{ds_popover}
\title{Popover Component}
\usage{
ds_popover(..., variant = NULL, class = NULL)
}
\arguments{
\item{...}{Popover content}

\item{variant}{Popover variant ("default", "tinted")}

\item{class}{Additional CSS classes}
}
\value{
A Shiny tag object
}
\description{
Create a popover using Designsystemet styles.
}
\examples{
ds_popover("Helpful information about this field.")
ds_popover("Highlighted note", variant = "tinted")
}
