% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-misc.R
\name{ds_skeleton}
\alias{ds_skeleton}
\title{Skeleton Component}
\usage{
ds_skeleton(variant = "text", width = NULL, height = NULL, class = NULL)
}
\arguments{
\item{variant}{Skeleton variant ("text", "circle", "rectangle")}

\item{width}{Width of the skeleton}

\item{height}{Height of the skeleton}

\item{class}{Additional CSS classes}
}
\value{
A Shiny tag object
}
\description{
Create a loading skeleton placeholder using Designsystemet styles.
}
\examples{
ds_skeleton(variant = "text", width = "200px")
ds_skeleton(variant = "circle", width = "50px", height = "50px")
}
