% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-misc.R
\name{ds_spinner}
\alias{ds_spinner}
\title{Spinner Component}
\usage{
ds_spinner(title = "Laster...", size = NULL, class = NULL)
}
\arguments{
\item{title}{Accessibility title for the spinner}

\item{size}{Size variant ("sm", "md", "lg")}

\item{class}{Additional CSS classes}
}
\value{
A Shiny tag object
}
\description{
Create a loading spinner using Designsystemet styles.
}
\examples{
ds_spinner(title = "Loading...")
}
