% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-tabs.R
\name{ds_tab}
\alias{ds_tab}
\title{Tab Button}
\usage{
ds_tab(label, value, selected = FALSE, ...)
}
\arguments{
\item{label}{The tab label text}

\item{value}{The value to return when this tab is selected}

\item{selected}{Whether this tab is initially selected}

\item{...}{Additional attributes}
}
\value{
A Shiny tag object
}
\description{
A single tab button within ds_tablist.
}
\examples{
ds_tab("Overview", value = "overview", selected = TRUE)
}
