% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-tabs.R
\name{ds_tablist}
\alias{ds_tablist}
\title{Tab List Container}
\usage{
ds_tablist(..., class = NULL)
}
\arguments{
\item{...}{Child elements (ds_tab elements)}

\item{class}{Additional CSS classes}
}
\value{
A Shiny tag object
}
\description{
Container for tab buttons within ds_tabs.
}
\examples{
ds_tablist(
  ds_tab("First", value = "tab1", selected = TRUE),
  ds_tab("Second", value = "tab2")
)
}
