% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-tabs.R
\name{ds_tabs}
\alias{ds_tabs}
\title{Tabs Component}
\usage{
ds_tabs(inputId, ..., class = NULL)
}
\arguments{
\item{inputId}{The input slot for Shiny reactivity}

\item{...}{Child elements (typically ds_tablist and ds_tabpanel elements)}

\item{class}{Additional CSS classes}
}
\value{
A Shiny tag object
}
\description{
Create a tabbed interface using Designsystemet tabs.
}
\examples{
ds_tabs("mytabs",
  ds_tablist(
    ds_tab("First", value = "tab1", selected = TRUE),
    ds_tab("Second", value = "tab2")
  ),
  ds_tabpanel(value = "tab1", "Content 1"),
  ds_tabpanel(value = "tab2", "Content 2")
)
}
