% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-table.R
\name{ds_tbody}
\alias{ds_tbody}
\title{Table Body}
\usage{
ds_tbody(...)
}
\arguments{
\item{...}{Body rows}
}
\value{
A Shiny tag object
}
\description{
Create a table body section.
}
\examples{
ds_tbody(ds_tr(ds_td("John"), ds_td("john@example.com")))
}
