% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-table.R
\name{ds_td}
\alias{ds_td}
\title{Table Data Cell}
\usage{
ds_td(...)
}
\arguments{
\item{...}{Cell content}
}
\value{
A Shiny tag object
}
\description{
Create a table data cell.
}
\examples{
ds_td("Cell content")
}
