% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-table.R
\name{ds_th}
\alias{ds_th}
\title{Table Header Cell}
\usage{
ds_th(..., scope = "col")
}
\arguments{
\item{...}{Cell content}

\item{scope}{Cell scope ("col", "row")}
}
\value{
A Shiny tag object
}
\description{
Create a table header cell.
}
