% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-table.R
\name{ds_thead}
\alias{ds_thead}
\title{Table Header}
\usage{
ds_thead(...)
}
\arguments{
\item{...}{Header rows}
}
\value{
A Shiny tag object
}
\description{
Create a table header section.
}
\examples{
ds_thead(ds_tr(ds_th("Name"), ds_th("Email")))
}
