% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-details.R
\name{ds_toggle_group}
\alias{ds_toggle_group}
\title{Toggle Group Component}
\usage{
ds_toggle_group(inputId, ..., size = NULL, class = NULL)
}
\arguments{
\item{inputId}{The input slot for Shiny reactivity}

\item{...}{Toggle buttons}

\item{size}{Size variant ("sm", "md", "lg")}

\item{class}{Additional CSS classes}
}
\value{
A Shiny tag object
}
\description{
Create a toggle button group using Designsystemet styles.
}
