% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-table.R
\name{ds_tr}
\alias{ds_tr}
\title{Table Row}
\usage{
ds_tr(...)
}
\arguments{
\item{...}{Row cells}
}
\value{
A Shiny tag object
}
\description{
Create a table row.
}
\examples{
ds_tr(ds_td("Cell 1"), ds_td("Cell 2"))
}
