% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-details.R
\name{hide_ds_dialog}
\alias{hide_ds_dialog}
\title{Close a dialog from the server}
\usage{
hide_ds_dialog(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{The \code{id} of the \code{ds_dialog()} to close.}

\item{session}{Shiny session object.}
}
\value{
Called for side effects.
}
\description{
Sends a message to the browser to call \code{close()} on a \code{ds_dialog()} element.
}
\examples{
\dontrun{
observeEvent(input$cancel_btn, {
  hide_ds_dialog("confirm-dialog")
})
}
}
