% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-details.R
\name{show_ds_dialog}
\alias{show_ds_dialog}
\title{Open a dialog from the server}
\usage{
show_ds_dialog(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{The \code{id} of the \code{ds_dialog()} to open.}

\item{session}{Shiny session object.}
}
\value{
Called for side effects.
}
\description{
Sends a message to the browser to call \code{showModal()} on a \code{ds_dialog()}
element already present in the UI.
}
\examples{
\dontrun{
observeEvent(input$open_btn, {
  show_ds_dialog("confirm-dialog")
})
}
}
