% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-input.R
\name{update_ds_checkbox}
\alias{update_ds_checkbox}
\title{Update Checkbox}
\usage{
update_ds_checkbox(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL
)
}
\arguments{
\item{session}{The Shiny session object}

\item{inputId}{The input ID}

\item{value}{The new checked state (TRUE/FALSE)}
}
\value{
Called for its side effect. Returns \code{NULL} invisibly.
}
\description{
Change the checked state from the server.
}
\examples{
\dontrun{
update_ds_checkbox(session, "agree", value = FALSE)
}
}
