% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-pagination.R
\name{update_ds_pagination}
\alias{update_ds_pagination}
\title{Update Pagination Input}
\usage{
update_ds_pagination(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  current = NULL,
  total = NULL
)
}
\arguments{
\item{session}{The Shiny session object}

\item{inputId}{The input ID of the pagination component}

\item{current}{The new current page number}

\item{total}{The new total number of pages}
}
\description{
Change the current page or total pages from the server.
}
