% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-input.R
\name{update_ds_radio_group}
\alias{update_ds_radio_group}
\title{Update Radio Group}
\usage{
update_ds_radio_group(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL
)
}
\arguments{
\item{session}{The Shiny session object}

\item{inputId}{The input ID}

\item{value}{The new selected value}
}
\value{
Called for its side effect. Returns \code{NULL} invisibly.
}
\description{
Change the selected value of a \code{\link[=ds_radio_group]{ds_radio_group()}} from the server.
}
\examples{
\dontrun{
update_ds_radio_group(session, "opts", value = "b")
}
}
