% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-input.R
\name{update_ds_select}
\alias{update_ds_select}
\title{Update Select}
\usage{
update_ds_select(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  choices = NULL
)
}
\arguments{
\item{session}{The Shiny session object}

\item{inputId}{The input ID}

\item{value}{The new selected value}

\item{choices}{New named vector of choices}
}
\value{
Called for its side effect. Returns \code{NULL} invisibly.
}
\description{
Change the selected value or choices from the server.
}
\examples{
\dontrun{
update_ds_select(session, "country", value = "no")
}
}
