% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-suggestion.R
\name{update_ds_suggestion}
\alias{update_ds_suggestion}
\title{Update Suggestion Input}
\usage{
update_ds_suggestion(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL,
  choices = NULL
)
}
\arguments{
\item{session}{The Shiny session object}

\item{inputId}{The input ID of the suggestion component}

\item{value}{The new value}

\item{choices}{New character vector of choices}
}
\value{
Called for its side effect. Returns \code{NULL} invisibly.
}
\description{
Change the value or choices from the server.
}
\examples{
\dontrun{
update_ds_suggestion(session, "fruit", value = "Apple")
}
}
