% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-misc.R
\name{ds_avatar_stack}
\alias{ds_avatar_stack}
\title{Avatar Stack Component}
\usage{
ds_avatar_stack(..., class = NULL)
}
\arguments{
\item{...}{Avatar elements}

\item{class}{Additional CSS classes}
}
\value{
A Shiny tag object
}
\description{
Create a stack of avatars using Designsystemet styles.
}
