% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-input.R
\name{ds_radio}
\alias{ds_radio}
\title{Radio Button Component}
\usage{
ds_radio(
  inputId,
  label,
  value,
  name,
  checked = FALSE,
  size = NULL,
  disabled = FALSE,
  ...
)
}
\arguments{
\item{inputId}{The input slot for Shiny reactivity}

\item{label}{The radio button label}

\item{value}{The value for this radio button}

\item{name}{The name grouping radio buttons together}

\item{checked}{If TRUE, this radio is initially selected}

\item{size}{Size variant ("sm", "md", "lg")}

\item{disabled}{If TRUE, disables the radio button}

\item{...}{Additional attributes}
}
\value{
A Shiny tag object
}
\description{
Create a styled radio button using Designsystemet styles.
}
