% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-details.R
\name{ds_search}
\alias{ds_search}
\title{Search Component}
\usage{
ds_search(
  inputId,
  value = "",
  placeholder = NULL,
  size = NULL,
  class = NULL,
  ...
)
}
\arguments{
\item{inputId}{The input slot for Shiny reactivity}

\item{value}{Initial value}

\item{placeholder}{Placeholder text}

\item{size}{Size variant ("sm", "md", "lg")}

\item{class}{Additional CSS classes}

\item{...}{Additional attributes}
}
\value{
A Shiny tag object
}
\description{
Create a search input using Designsystemet styles.
}
\examples{
ds_search("search", placeholder = "Search…")
}
