% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-tabs.R
\name{ds_tabpanel}
\alias{ds_tabpanel}
\title{Tab Panel}
\usage{
ds_tabpanel(..., value, class = NULL)
}
\arguments{
\item{...}{Panel content}

\item{value}{The value that matches the corresponding ds_tab}

\item{class}{Additional CSS classes}
}
\value{
A Shiny tag object
}
\description{
Content panel for a tab.
}
\examples{
ds_tabpanel(value = "overview", ds_paragraph("Content here."))
}
