% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-tooltip.R
\name{ds_tooltip}
\alias{ds_tooltip}
\title{Tooltip}
\usage{
ds_tooltip(element, text, placement = "top")
}
\arguments{
\item{element}{The element to attach the tooltip to}

\item{text}{The tooltip text}

\item{placement}{Tooltip placement ("top", "bottom", "left", "right")}
}
\value{
The element with tooltip attributes added
}
\description{
Add a tooltip to an element using the data-tooltip attribute.
}
\examples{
ds_tooltip(
  ds_button("Hover me"),
  text = "This is helpful information"
)
}
