% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-input.R
\name{update_ds_input}
\alias{update_ds_input}
\title{Update Text Input}
\usage{
update_ds_input(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  value = NULL
)
}
\arguments{
\item{session}{The Shiny session object}

\item{inputId}{The input ID}

\item{value}{The new value}
}
\value{
Called for its side effect. Returns \code{NULL} invisibly.
}
\description{
Change the value from the server.
}
\examples{
\dontrun{
update_ds_input(session, "name", value = "")
}
}
