% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-tabs.R
\name{update_ds_tabs}
\alias{update_ds_tabs}
\title{Update Tabs Input}
\usage{
update_ds_tabs(
  session = shiny::getDefaultReactiveDomain(),
  inputId,
  selected = NULL
)
}
\arguments{
\item{session}{The Shiny session object}

\item{inputId}{The input ID of the tabs component}

\item{selected}{The value of the tab to select}
}
\description{
Change the selected tab from the server.
}
